+++ src/assets/index.tsx
@@ -1,0 +1,6 @@
+import HeartOn from '@/assets/icons/heart_on.svg?react';
+import HomeIcon from '@/assets/icons/home.svg?react';
+import SearchIcon from '@/assets/icons/search.svg?react';
+import MyIcon from '@/assets/icons/my.svg?react';
+
+export { HeartOn, HomeIcon, SearchIcon, MyIcon };
