+++ src/components/common/ReviewCard/ReviewCard.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import HeartOn from '@/assets/icons/heart_on.svg?react';
+import { HeartOn } from '@/assets';
 interface ReviewCardProps {
   imageUrl: string;
   tags: string[];
