+++ src/components/index.tsx
@@ -1,4 +1,6 @@
 import Button from '@/components/common/Button';
 import BottomNavigation from '@/components/common/BottomNavigation/BottomNavigation';
+import BestCinemaCard from '@/components/common/BestCinemaCard/BestCinemaCard';
+import ReviewCard from '@/components/common/ReviewCard/ReviewCard';
 
-export { Button, BottomNavigation };
+export { Button, BottomNavigation, BestCinemaCard, ReviewCard };
